# openapi.yaml
# -- Code Review --
#
#
# - Missing `description` in `info` block
#
# - Missing `servers` block:
#     servers:
#       - url: https://messages-api.com
#
# - Missing `tags` block:
#     tags:
#       - name: message
#         description: Message operations
#
# - Missing error responses
#
# - Missing `components` block that can be reusable
#      components:
#        schemas:
#          Message:
#            type: object
#            properties:
#              uuid:
#                type: string
#                format: uuid
#                description: Unique identifier for the message
#                example: "1f00bd14-9296-6fd6-8947-6fa761676088"
#              text:
#                type: string
#                description: The message content
#                example: "Hello, World!"
#              status:
#                type: string
#                description: The current status of the message
#                enum:
#                  - sent
#                  - read
#                example: "sent"
#          Error:
#            type: object
#            properties:
#              error:
#                type: string
#                description: Error message
#                example: "Text is required"
#
# - Missing `security` block
#

openapi: 3.0.0
info:
  title: Messages API
  version: 1.0.0
paths:
  /messages:
    get:
      summary: Get Messages
      description: Retrieves a list of messages.
      parameters:
        - name: status
          in: query
          description: Filter messages by status
          required: false
          schema:
            type: "string"
            enum:
              - "sent"
              - "read"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  messages:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        uuid: 
                          type: "string"
                        text:
                          type: "string"
                        status:
                          type: "string"
                          enum:
                            - "sent"
                            - "read"
              example:
                messages:
                  - text: "Hello, World!"
                    status: "read"
                  - text: "How are you today?"
                    status: "sent"
                  - text: "Thanks, I'm fine. How about you?"
                    status: "sent"
  /messages/send:
    get:
      summary: Send a Message
      parameters:
        - name: text
          in: query
          description: The text you want to send
          required: true
          schema:
            type: "string"
      responses:
        '204':
          description: Successfully sent